Update.
2003-05-01  Ulrich Drepper  <drepper@redhat.com>

	* malloc/malloc.c (mSTATs): Call ptmalloc_init if necessary.

2003-04-29  Andreas Schwab  <schwab@suse.de>

	* string/test-strcat.c (do_one_test): Fix attribute name.

+++ nptl/sysdeps/unix/sysv/linux/s390/bits/pthreadtypes.h
@@ -94,7 +94,7 @@ typedef union
     unsigned long long int __woken_seq;
   } __data;
   char __size[__SIZEOF_PTHREAD_COND_T];
-  long int __align;
+  long long int __align;
 } pthread_cond_t;
 
 typedef union
@@ -118,7 +118,7 @@ typedef int pthread_once_t;
 typedef union
 {
 # if __WORDSIZE == 64
-struct
+  struct
   {
     int __lock;
     unsigned int __nr_readers;
